Chore: assorted cleanup and bugs
Hello! I'm the website officer at UNDRR. As you know we've been looking at your work for inspiration.

As part of that, we've noticed a few little things and I thought it might be nice to share them back.

In this PR:

- The Readme.md says Yarn should be used, so I've made docs more clearly suggest using Yarn
  - It seems odd there is a `package-lock.json` and a `yarn.lock`, but I didn't delete the `package-lock.json` in this PR
- A few assorted spelling things
- Updates stylelintrc.json to not do unhelpful things
- Fixes a Case typo on the Chips import

Let me know if this is helpful and we'll try to contribute back anything else we come across.

And, finally, thanks for being supportive of UNDRR's effort to roll our own library!