#970 - Multi select event trigger cleanup
Added the "select_trigger_id" and "select_trigger_data_id" to the multiSelectToggle event. This grabs an identifier on the triggering button that toggles the select element.
The "target" is the the main select item that is triggered by a select button.

select.min.js is from very much the same, and only contains linting/prettify changes that were applied to the select.js file.